New architecture
I simply split "rsdl.h" into a bunch of separate modules in "include/". Now "rsdl.h" is just a wrapper that bundles everything together for the laziest users

+++ src/utils.cpp
@@ -1,4 +1,4 @@
-#include "rsdl.hpp"
+#include "include/all_utils.h"
 #include "SDL3/SDL_misc.h"
 
 namespace rsdl{
@@ -7,8 +7,8 @@ void openURL(const char* url){
     if (!url) return;
     SDL_OpenURL(url);
 }
-uint64_t getGarbage(){
-    uint64_t garbage;
+int getGarbage(){
+    int garbage;
     return garbage;
 }
